Tethys
The 2007 OCG text does specify that it needs to be face-up, and the 2011 TCG errata is therefore likely a clarification and not a functional change. 
There's also nothing i'm aware of stating that it received any functional change nor was ruled differently at the time.

Adding for all formats it's legal in - clear that it never deviated from current errata.